requirements.txt
black==22.1.0
boto3==1.20.53
Django==3.2
django-htmx==1.8.0
django-import-export==2.7.1
django-smart-selects==1.5.9
django-storages==1.12.3
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
uWSGI==2.0.19.1
xhtml2pdf==0.2.5
reportlab==3.6.6

